requirements.txt
pyperclip
shapely
boltons
svg.path
pycodestyle
